requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.0.4
django-cors-headers==3.14.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
pep8-naming==0.13.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
python-dotenv==1.0.0
